% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc.R
\name{calc_elc}
\alias{calc_elc}
\title{Calculate the expected loss}
\usage{
calc_elc(wtp, psa)
}
\arguments{
\item{wtp}{vector of willingness to pay thresholds}

\item{psa}{object of class \code{psa}, produced by function
\code{\link{make_psa_obj}}}
}
\value{
object with classes \code{elc} and \code{data.frame}
}
\description{
Calculate the expected loss
}
