% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi.R
\name{calc_evpi}
\alias{calc_evpi}
\title{Expected Value of Perfect Information (EVPI)}
\usage{
calc_evpi(wtp, psa, pop = 1)
}
\arguments{
\item{wtp}{Numeric vector with willingness-to-pay (WTP) thresholds}

\item{psa}{psa object from \code{\link{make_psa_obj}}}

\item{pop}{A scalar that corresponds to the total population}
}
\value{
evpi A data frame with the EVPI at each WTP threshold.
}
\description{
\code{calc_evpi} is used to compute the expected value of perfect information
(EVPI) from a probabilistic sensitivity analysis (PSA) dataset.
}
\section{Details}{

\code{evpi} calculates the value of eliminating all the uncertainty of a
cost-effectiveness analysis at each WTP threshold.
}

\keyword{benefit}
\keyword{expected}
\keyword{information;}
\keyword{monetary}
\keyword{net}
\keyword{of}
\keyword{perfect}
\keyword{value}
