% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_loss.R
\name{calc_exp_loss}
\alias{calc_exp_loss}
\title{Calculate the expected loss at a range of willingness-to-pay thresholds}
\usage{
calc_exp_loss(wtp, psa)
}
\arguments{
\item{wtp}{vector of willingness to pay thresholds}

\item{psa}{object of class \code{psa}, produced by function
\code{\link{make_psa_obj}}}
}
\value{
object with classes \code{exp_loss} and \code{data.frame}
}
\description{
The expected loss is the ...
}
\details{
Visualize the expected loss at a variety of WTP thresholds using \code{\link{plot.exp_loss}}.
}
\examples{
data("example_psa_obj")
wtp <- seq(1e4, 1e5, by = 1e4)
exp_loss <- calc_exp_loss(wtp, example_psa_obj)

# can use head(), summary(), print(), etc.
head(exp_loss)

# plot an expected loss curve (ELC)
plot(exp_loss)

# the y axis is on a log scale by default
plot(exp_loss, log_y = FALSE)

}
\seealso{
\code{\link{plot.exp_loss}}, \code{\link{make_psa_obj}}
}
