% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{ceac}
\alias{ceac}
\title{Cost-Effectiveness Acceptability Curves (CEAC)}
\usage{
ceac(wtp, outcomes, strategies = NULL)
}
\arguments{
\item{wtp}{Numeric vector with willingness-to-pay (WTP) thresholds}

\item{outcomes}{Matrix with the model outputs. The outcomes must be ordered
in such a way that for each strategy the cost must appear first then the
effectiveness.}

\item{strategies}{String vector with the name of the strategies}
}
\value{
ceac A melted data frame with each strategy's probability of being
cost-effective at each WTP threshold.
}
\description{
\code{ceac} is used to compute and plot the cost-effectiveness acceptability
curves (CEAC) from a probabilistic sensitivity analysis (PSA) dataset.
}
\section{Details}{

\code{ceac} computes the probability of each of the strategies being
cost-effective at each \code{wtp} threshold.
}

\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curves}
