% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icers.R
\name{compute_icers}
\alias{compute_icers}
\title{compute icers for non-dominated strategies}
\usage{
compute_icers(non_d)
}
\arguments{
\item{non_d}{a data frame of non-dominated strategies, with columns
"Strategy", "Cost", and "Effect"}
}
\value{
the input dataframe with columns "Inc_Cost",
"Inc_Effect", and "ICER" appended
}
\description{
compute icers for non-dominated strategies
}
\keyword{internal}
