% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_oneway.R
\name{create_dsa_oneway}
\alias{create_dsa_oneway}
\title{create deterministic sensitivity analysis object}
\usage{
create_dsa_oneway(params, effectiveness, strategies, cost = NULL,
  currency = "$")
}
\arguments{
\item{params}{parameter values associated with}
}
\description{
create deterministic sensitivity analysis object
}
