% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_twoway.R
\name{create_dsa_twoway}
\alias{create_dsa_twoway}
\title{create two-way deterministic sensitivity analysis object}
\usage{
create_dsa_twoway(params, effectiveness, strategies, cost = NULL,
  currency = "$")
}
\arguments{
\item{params}{parameter values associated with}
}
\description{
create two-way deterministic sensitivity analysis object
}
