% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sa.R
\name{create_sa}
\alias{create_sa}
\title{An generic sensitivity analysis object}
\usage{
create_sa(parameters, parnames, effectiveness, strategies, cost, currency)
}
\arguments{
\item{parameters}{Data frame with values for each simulation (rows) and parameter (columns).
The column names should be the parameter names.}

\item{effectiveness}{Data frame with the effectiveness for each simulation (rows) and strategy (columns)}

\item{strategies}{String vector with the name of the strategies}

\item{cost}{Data frame with the cost for each simulation (rows) and strategy (columns).}

\item{currency}{symbol for the currency being used (ex. "$", "£")}
}
\description{
This object is called by \code{\link{create_psa_obj}}
and \code{\link{create_dsa_oneway}}, and checks the structure of
each of the arguments before creating an SA object.
}
