% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example_psa.R
\docType{data}
\name{example_psa}
\alias{example_psa}
\title{Sample PSA data for testing}
\format{An object of class \code{list} of length 4.}
\usage{
example_psa
}
\description{
A dataset containing a number of PSA samples
}
\keyword{datasets}
