% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example_psa_obj.R
\docType{data}
\name{example_psa_obj}
\alias{example_psa_obj}
\title{Sample PSA data for testing}
\format{
An object of class \code{psa} of length 8.
}
\usage{
example_psa_obj
}
\description{
A psa object created from the data in example_psa
}
\keyword{datasets}
