% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ce_frontier.R
\name{get_ce_frontier}
\alias{get_ce_frontier}
\title{Find the CEA frontier, up to a given WTP level, by
identifying strategies with the highest NMB}
\usage{
get_ce_frontier(ce_df, max_wtp = Inf)
}
\description{
Find the CEA frontier, up to a given WTP level, by
identifying strategies with the highest NMB
}
\details{
Originally written by Sze Suen on Feb 25, 2015.
Modified by Fernando Alarid-Escudero and Caleb Easterly.

If code is used, please cite the below paper:

Suen S-C, Goldhaber-Fiebert JD. An Efficient,
}
