% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier-function.R
\name{getFrontier}
\alias{getFrontier}
\title{R functions to compute and plot CE Frontier ####}
\usage{
getFrontier(CEmat, maxWTP = Inf, plot = TRUE)
}
\description{
R functions to compute and plot CE Frontier ####
}
