% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.ceac.R
\name{head.ceac}
\alias{head.ceac}
\title{head method for the ceac object}
\usage{
\method{head}{ceac}(x, n = 6L, ...)
}
\arguments{
\item{x}{object returned from the \code{ceac} function}

\item{n}{number of lines to display}

\item{...}{further arguments (not used)}
}
\description{
head method for the ceac object
}
