% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hund_strat.R
\docType{data}
\name{hund_strat}
\alias{hund_strat}
\title{Sample deterministic data for testing}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
hund_strat
}
\description{
A dataset containing 100 strategies
}
\keyword{datasets}
