% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{is_owsa}
\alias{is_owsa}
\title{check that object is owsa object}
\usage{
is_owsa(obj)
}
\arguments{
\item{obj}{the object to be checked for the owsa class}
}
\value{
returns TRUE if object is "owsa" object and FALSE if not.
}
\description{
check that object is owsa object
}
\keyword{internal}
