% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{make_param_seq}
\alias{make_param_seq}
\title{make a parameter sequence}
\usage{
make_param_seq(p, ranges, nsamp, psa_paramvals)
}
\arguments{
\item{p}{parameter of interest}

\item{ranges}{named vector of parameter ranges}

\item{nsamp}{number of points from the range}

\item{psa_paramvals}{sampled values from the PSA. used to calculate the
range if none is supplied}
}
\description{
make a parameter sequence
}
\keyword{internal}
