% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{make_psa_obj}
\alias{make_psa_obj}
\title{An object to hold PSA results.}
\usage{
make_psa_obj(cost, effectiveness, strategies = NULL, currency = "$")
}
\arguments{
\item{cost}{Matrix with the cost for each simulation (rows) and strategy (columns).}

\item{effectiveness}{Matrix with the effectiveness for each simulation (rows) and strategy (columns)}

\item{strategies}{String vector with the name of the strategies}

\item{currency}{symbol for the currency being used (ex. "$", "£")}
}
\description{
An object to hold PSA results.
}
