% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{metamod}
\alias{metamod}
\title{linear regression metamodeling}
\usage{
metamod(analysis = c("oneway", "twoway"), psa, parms = NULL,
  strategies = NULL, outcome = c("eff", "cost", "nhb", "nmb"),
  wtp = NULL, type = "poly", poly.order = 2)
}
\arguments{
\item{analysis}{either "oneway" or "twoway"}

\item{psa}{psa object}

\item{parms}{String with the name of the parameter of interest}

\item{strategies}{vector of strategies to consider. The default (NULL) is that all strategies are considered.}

\item{outcome}{either effectiveness ("eff"), cost, net health benefit ("nhb"), or net monetary benefit ("nmb")}

\item{wtp}{if outcome is NHB or NMB, must provide the willingness-to-pay threshold}

\item{type}{type of metamodel}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\description{
linear regression metamodeling
}
