% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{metamodel}
\alias{metamodel}
\title{Linear regression metamodeling}
\usage{
metamodel(analysis = c("oneway", "twoway"), psa, parms = NULL,
  strategies = NULL, outcome = c("eff", "cost", "nhb", "nmb",
  "nhb_loss", "nmb_loss"), wtp = NULL, type = c("linear", "gam",
  "poly"), poly.order = 2, k = -1)
}
\arguments{
\item{analysis}{either "oneway" or "twoway"}

\item{psa}{psa object}

\item{parms}{String with the name of the parameter of interest}

\item{strategies}{vector of strategies to consider. The default (NULL) is that all strategies are considered.}

\item{outcome}{either effectiveness ("eff"), cost, net health benefit ("nhb"), or net monetary benefit ("nmb")}

\item{wtp}{if outcome is NHB or NMB, must provide the willingness-to-pay threshold}

\item{type}{type of metamodel}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}

\item{k}{ the dimension of the basis used to represent the smooth term.
    The default depends on the number of variables that the smooth is a
    function of. \code{k} should not be less than the dimension of the
    null space of the penalty for the term (see
    \code{\link{null.space.dimension}}), but will be reset if
  it is. See \code{\link{choose.k}} for further information.}
}
\value{
A metamodel object, which contains a list of metamodels and other relevant information.
}
\description{
This function estimates a linear regression metamodel for
a given decision-analytic model by using the results of a probabilistic sensitivity analysis (PSA)
}
\details{
The most important option is \code{analysis}, which can be either \code{"oneway"}
or \code{twoway}. If \code{analysis == "oneway"}, a separate metamodel is created
for each combination of the parameters in \code{parms} and strategies in \code{strategies}
(by default, this is all strategies and parameters).

If \code{analysis == "twoway"}, \code{parms} must be a vector of two parameters, and a metamodel
is created with these two parameters for each strategy in \code{strategies}.
}
\seealso{
\code{\link{predict.metamodel}},
\code{\link{make_psa_obj}},
\code{\link{owsa}},
\code{\link{twsa}}
}
