% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{mm_run_reg}
\alias{mm_run_reg}
\title{this should have just one parm of int
when the analysis is one-way}
\usage{
mm_run_reg(dep, parms_of_int, dat, all_parms, type, poly.order)
}
\description{
this should have just one parm of int
when the analysis is one-way
}
