% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_common_aes.R
\name{number_ticks}
\alias{number_ticks}
\title{Number of ticks for \code{ggplot2} plots}
\usage{
number_ticks(n)
}
\arguments{
\item{n}{integer giving the desired number of ticks on axis of
\code{ggplot2} plots. Non-integer values are rounded down.}
}
\value{
a vector of axis-label breaks
}
\description{
Function for determining number of ticks on axis of \code{ggplot2} plots.
}
\section{Details}{

Based on function \code{pretty}.
}

\keyword{internal}
