% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{offset_trans}
\alias{offset_trans}
\title{transformation for owsa_tornado}
\usage{
offset_trans(offset = 0)
}
\arguments{
\item{offset}{the offset for the transformation (kind of the new 0)}
}
\description{
transformation for owsa_tornado
}
\keyword{internal}
