% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic-SA.R
\name{owsa_det_wrapper_of_user_model}
\alias{owsa_det_wrapper_of_user_model}
\title{Specifically used for owsa_det}
\usage{
owsa_det_wrapper_of_user_model(x, user_fun, parm_name, tmp_input,
  tmp_replace)
}
\description{
Specifically used for owsa_det
}
