% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi.R
\name{plot.evpi}
\alias{plot.evpi}
\title{Plot of Expected Value of Perfect Information (EVPI)}
\usage{
\method{plot}{evpi}(x, title = "Expected Value of Perfect Information",
  txtsize = 12, currency = "$", effect_units = "QALY", ...)
}
\arguments{
\item{x}{object of class \code{evpi}, produced by function
\code{\link{calc_evpi}}}

\item{title}{String with graph's title}

\item{txtsize}{number with text size}

\item{currency}{String with currency used in the cost-effectiveness analysis (CEA).
Default: $, but it could be any currency symbol or word (e.g., £, €, peso)}

\item{effect_units}{Units of effectiveness. Default: QALY}

\item{...}{further arguments to plot() (not used)}
}
\value{
A \code{ggplot2} object with the EVPI
}
\description{
Plots the EVPI as a \code{ggplot2} object calculated with \code{\link{calc_evpi}}.
}
\keyword{expected}
\keyword{information}
\keyword{of}
\keyword{perfect}
\keyword{value}
