% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icers.R
\name{plot.icers}
\alias{plot.icers}
\title{Plot of ICERs}
\usage{
\method{plot}{icers}(x, txtsize = 12, currency = "$",
  effect_units = "QALYs", label = c("frontier", "all", "none"),
  label_max_char = 8, plot_frontier_only = FALSE, alpha = 1,
  n_x_ticks = 6, n_y_ticks = 6, ...)
}
\arguments{
\item{x}{Object of class \code{ceac}. A melted data frame produced by
function \code{ceac} with each strategy's probability of being
cost-effective for each willingness-to-pay (WTP) threshold}

\item{txtsize}{base text size}

\item{currency}{string. with currency used in the cost-effectiveness analysis (CEA).}

\item{effect_units}{string. unit of effectiveness}

\item{label}{whether to label strategies on the efficient frontier, all strategies, or none.
defaults to frontier.}

\item{label_max_char}{max number of characters to label the strategies - longer strategies will be
truncated to save space.}

\item{plot_frontier_only}{only plot the efficient frontier}

\item{alpha}{opacity of points}

\item{n_x_ticks}{number of axis ticks}

\item{n_y_ticks}{number of axis ticks}

\item{...}{further arguments to plot (not used)}
}
\description{
Plots the CEAC as a \code{ggplot2} object calculated with \code{\link{ceac}}.
}
