% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{predict.metamodel}
\alias{predict.metamodel}
\title{Predict from a metamodel}
\usage{
\method{predict}{metamodel}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{object with class "metamodel"}

\item{newdata}{values for parameter of interest}

\item{...}{further arguments to \code{predict} (not used)}
}
\description{
Predict from a metamodel
}
