% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{predict_ga}
\alias{predict_ga}
\title{Function to compute the preposterior for each of the
basis functions of the GAM model.}
\usage{
predict_ga(object, n, n0)
}
\arguments{
\item{object}{gam object}

\item{n}{scalar or vector of new sample size to compute evsi on}

\item{n0}{scalar or vector of effective prior sample size}
}
\description{
Function to compute the preposterior for each of the
basis functions of the GAM model.
}
\keyword{internal}
