% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{predict_matrix_tensor_smooth_ga}
\alias{predict_matrix_tensor_smooth_ga}
\title{Predict matrix tensor smooth (GA)}
\usage{
predict_matrix_tensor_smooth_ga(object, param_vals, vrf = rep(1,
  ncol(param_vals)))
}
\description{
Function to compute the preposterior for each of the
basis functions for one or more parameters and calculates
the tensor product if more than one parameter is selected
(Heavily based on function Predict.matrix.tensor.smooth from
mgcv package)
}
\keyword{internal}
