% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{predict_smooth_ga}
\alias{predict_smooth_ga}
\title{Function to compute the preposterior for each of the
basis functions of a smooth for one parameter}
\usage{
predict_smooth_ga(object, param_vals, vrf = 1)
}
\description{
Function to compute the preposterior for each of the
basis functions of a smooth for one parameter
}
\keyword{internal}
