% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{print.ceac}
\alias{print.ceac}
\title{print method for the ceac object}
\usage{
\method{print}{ceac}(x, ...)
}
\arguments{
\item{x}{object returned from the \code{ceac} function}

\item{...}{further arguments (not used)}
}
\description{
print method for the ceac object
}
