% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sa.R
\name{print.sa}
\alias{print.sa}
\title{print a psa object}
\usage{
\method{print}{sa}(x, all_strat = FALSE, ...)
}
\arguments{
\item{x}{the psa object}

\item{all_strat}{whether or not to print the full list of strategies. defaults to FALSE, which truncates
the strategy list to 5}

\item{...}{further arguments to print (not used)}
}
\value{
None (invisible NULL).
}
\description{
print a psa object
}
