% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_data.R
\docType{data}
\name{psa}
\alias{psa}
\title{Sample PSA data for testing}
\format{An object of class \code{data.frame} with 10000 rows and 15 columns.}
\usage{
psa
}
\description{
A dataset containing a number of PSA samples
}
\keyword{datasets}
