% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cdiff.R
\docType{data}
\name{psa_cdiff}
\alias{psa_cdiff}
\title{Sample PSA dataset}
\format{
An object of class \code{psa} (inherits from \code{sa}) of length 6.
}
\usage{
psa_cdiff
}
\description{
Sample PSA dataset
}
\keyword{datasets}
