% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa_results}
\alias{psa_results}
\title{An object to hold PSA results.}
\usage{
psa_results(cost, effectiveness, strategies = NULL)
}
\arguments{
\item{cost}{Matrix with the cost for each simulation (rows) and strategy (columns).}

\item{effectiveness}{Matrix with the effectiveness for each simulation (rows) and strategy (columns)}

\item{strategies}{String vector with the name of the strategies}
}
\description{
An object to hold PSA results.
}
