% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_psa_samp.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random number generation for the dirichlet distribution with parameter vector alpha.}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of observations}

\item{alpha}{vector of parameters defining dirichlet distribution


 @importFrom stats rgamma}
}
\description{
Random number generation for the dirichlet distribution with parameter vector alpha.
}
