% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{summary.metamodel}
\alias{summary.metamodel}
\title{Summary of metamodel}
\usage{
\method{summary}{metamodel}(object, ...)
}
\arguments{
\item{object}{metamodel to summarize}

\item{...}{further arguments to summary}
}
\value{
a \code{data.frame} containing the r-squared for each strategy and parameter's metamodel.
}
\description{
Summary of metamodel
}
