% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{summary.psa}
\alias{summary.psa}
\title{summarize a psa object across all simulations}
\usage{
\method{summary}{psa}(object, calc_sds = FALSE, ...)
}
\arguments{
\item{object}{the psa object}

\item{calc_sds}{whether or not to calculate the standard deviations. Defaults to FALSE}

\item{...}{further arguments to summary (not used)}
}
\value{
a \code{data.frame} containing the mean cost and effectiveness for each strategy and, if requested,
the standard deviations of the cost and effectiveness for each strategy.
}
\description{
summarize a psa object across all simulations
}
