% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa.R
\name{twsa}
\alias{twsa}
\title{Two-way sensitivity analysis using linear regression metamodeling}
\usage{
twsa(psa, parm1, parm2, ranges = NULL, outcome = c("eff", "cost",
  "nhb", "nmb"), wtp = NULL, strategies = NULL, poly.order = 2)
}
\arguments{
\item{psa}{psa object}

\item{parm1}{String with the name of the first parameter of interest}

\item{parm2}{String with the name of the second parameter of interest}

\item{outcome}{either effectiveness ("eff"), cost, net health benefit ("nhb"), or net monetary benefit ("nmb")}

\item{wtp}{if outcome is NHB or NMB, must provide the willingness-to-pay threshold}

\item{strategies}{String vector with the name of the strategies}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\value{
twsa A \code{ggplot2} object with the TWSA graph of \code{parm1} and
\code{parm2} on the outcome of interest that can be posteriorly formatted
with \code{ggplot2} function
}
\description{
This function displays a two-way sensitivity analysis (TWSA) graph
by estimating a linear regression metamodel of a PSA for a given
decision-analytic model
}
\keyword{analysis;}
\keyword{linear}
\keyword{metamodel}
\keyword{regression}
\keyword{sensitivity}
\keyword{two-way}
