% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic-SA.R
\name{twsa_det_wrapper_of_user_model}
\alias{twsa_det_wrapper_of_user_model}
\title{Specifically used for twsa_det}
\usage{
twsa_det_wrapper_of_user_model(x, user_fun, parm_name, tmp_input,
  tmp_replace)
}
\description{
Specifically used for twsa_det
}
