% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dsa.R
\name{wrapper_of_user_model}
\alias{wrapper_of_user_model}
\title{Wrapper function for owsa_det and twsa_det}
\usage{
wrapper_of_user_model(
  x,
  user_fun,
  param_name,
  tmp_input,
  tmp_replace,
  progress_bar = NULL,
  param_counter = NULL,
  nsamp = NULL
)
}
\arguments{
\item{x}{iteration}

\item{user_fun}{\code{FUN} input from users}

\item{param_name}{user-defined list of parameters of interest}

\item{tmp_input}{basecase values}

\item{tmp_replace}{values from predetermined DSA samples that will replace some values in \code{tmp_input}}
}
\value{
output of function defined by \code{user_fun}
}
\description{
Wrapper function for owsa_det and twsa_det
}
\keyword{internal}
