% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{calc_evsi}
\alias{calc_evsi}
\title{Calculate Expected Value of Sample Information}
\usage{
calc_evsi(
  psa,
  wtp,
  params = NULL,
  outcome = c("nhb", "nmb"),
  type = c("gam", "poly"),
  poly.order = 2,
  k = -1,
  n = 100,
  n0 = 10,
  pop = 1
)
}
\arguments{
\item{psa}{object of class psa, produced by \code{\link{make_psa_obj}}}

\item{wtp}{willingness-to-pay threshold}

\item{params}{A vector of parameter names to be analyzed in terms of EVPPI.}

\item{outcome}{either net monetary benefit (\code{"nmb"})
or net health benefit (\code{"nhb"})}

\item{type}{either generalized additive models (\code{"gam"}) or
polynomial models (\code{"poly"})}

\item{poly.order}{order of the polynomial, if \code{type == "poly"}}

\item{k}{basis dimension, if \code{type == "gam"}}

\item{n}{additional sample size}

\item{n0}{initial sample size}

\item{pop}{A scalar that corresponds to the total population}
}
\description{
Calculate Expected Value of Sample Information
}
