% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{calc_evsi}
\alias{calc_evsi}
\title{Calculate Expected Value of Sample Information}
\usage{
calc_evsi(psa, wtp, params = NULL, outcome = c("nhb_loss", "nmb_loss"),
  type = c("gam", "poly"), poly.order = 2, k = NA, n = 100,
  n0 = 10)
}
\arguments{
\item{psa}{object of class psa, produced by \code{\link{make_psa_obj}}}

\item{wtp}{willingness-to-pay threshold}

\item{params}{A vector of parameters for which to analyze the EVPPI.}

\item{outcome}{either net monetary benefit (\code{"nmb_loss"})
or net health benefit (\code{"nhb_loss"})}

\item{type}{either generalized additive models (\code{"gam"}) or
polynomial models (\code{"poly"})}

\item{poly.order}{order of the polynomial, if \code{type == "poly"}}

\item{k}{basis dimension, if \code{type == "gam"}}

\item{n}{additional sample size}

\item{n0}{initial sample size}
}
\description{
Calculate Expected Value of Sample Information
}
