% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_icers.R
\name{calculate_icers}
\alias{calculate_icers}
\title{function to calculate ICERS}
\usage{
calculate_icers(cost, effect, strategies)
}
\arguments{
\item{cost}{vector of cost for each strategy}

\item{effect}{vector of effect for each strategy}

\item{strategies}{character vector of strategy names}
}
\description{
function to calculate ICERS
}
