% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outcome.R
\name{calculate_outcome}
\alias{calculate_outcome}
\title{A function that is used to calculate all outcomes}
\usage{
calculate_outcome(outcome = c("nhb", "nmb", "eff", "cost", "nhb_loss",
  "nmb_loss"), cost, effect, wtp)
}
\arguments{
\item{outcome}{choice of outcome}

\item{cost}{data frame with costs}

\item{effect}{data frame with effects}

\item{wtp}{willingness-to-pay threshold}
}
\description{
A function that is used to calculate all outcomes
}
\keyword{internal}
