% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{ceac}
\alias{ceac}
\title{Cost-Effectiveness Acceptability Curves (CEAC)}
\usage{
ceac(wtp, costs, effectiveness, strategies = NULL)
}
\arguments{
\item{wtp}{Numeric vector with willingness-to-pay (WTP) thresholds}

\item{costs}{Matrix with the costs for each simulation (rows) and strategy (columns).}

\item{effectiveness}{Matrix with the effectiveness for each simulation (rows) and strategy (columns)}

\item{strategies}{String vector with the name of the strategies}
}
\value{
ceac A melted data frame with each strategy's probability of being
cost-effective at each WTP threshold.
}
\description{
\code{ceac} is used to compute and plot the cost-effectiveness acceptability
curves (CEAC) from a probabilistic sensitivity analysis (PSA) dataset.
}
\section{Details}{

\code{ceac} computes the probability of each of the strategies being
cost-effective (optimal?) at each \code{wtp} threshold.
}

\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curves}
