% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{ceac}
\alias{ceac}
\title{Cost-Effectiveness Acceptability Curves (CEAC)}
\usage{
ceac(wtp, psa)
}
\arguments{
\item{wtp}{Numeric vector with willingness-to-pay (WTP) thresholds}

\item{psa}{A psa object from \code{psa}}
}
\value{
ceac A melted data frame with each strategy's probability of being
cost-effective at each WTP threshold.
}
\description{
\code{ceac} is used to compute and plot the cost-effectiveness acceptability
curves (CEAC) from a probabilistic sensitivity analysis (PSA) dataset.
}
\section{Details}{

\code{ceac} computes the probability of each of the strategies being
cost-effective (optimal?) at each \code{wtp} threshold.
}

\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curves}
