% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_oneway.R
\name{create_dsa_oneway}
\alias{create_dsa_oneway}
\title{Create deterministic sensitivity analysis object}
\usage{
create_dsa_oneway(parameters, effectiveness, strategies, cost = NULL,
  currency = "$")
}
\arguments{
\item{parameters}{parameter values associated with effectiveness and outcomes.
The table must have two columns, with each parameter name in the first column
and the associated parameter value in the second column:

\tabular{ll}{
parameter      \tab value \cr
parm1 name     \tab parm1 val1 \cr
...            \tab ... \cr
parm2 name     \tab  parm2 val1 \cr
...            \tab ... \cr
}}

\item{effectiveness}{Matrix with the effectiveness for each simulation (rows) and strategy (columns)}

\item{strategies}{String vector with the name of the strategies}

\item{cost}{Matrix with the cost for each simulation (rows) and strategy (columns).}

\item{currency}{symbol for the currency being used (ex. "$", "£")}
}
\description{
Create deterministic sensitivity analysis object
}
