% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_twoway.R
\name{create_dsa_twoway}
\alias{create_dsa_twoway}
\title{Create one-way deterministic sensitivity analysis object}
\usage{
create_dsa_twoway(parameters, effectiveness, strategies, cost = NULL,
  currency = "$")
}
\arguments{
\item{parameters}{parameter values associated with effectiveness and outcomes.
The table must have two columns, one for each parameter.
The parameter names must be the column names.

\tabular{ll}{
parm1 name     \tab parm2 name \cr
parm1 val1     \tab parm2 val1 \cr
parm1 val2     \tab parm2 val2 \cr
...            \tab ... \cr
}}

\item{effectiveness}{Data frames containing costs and effectiveness data, respectively.
Each simulation should be a row of the data frame, and each strategy should be a column.
Naming the columns of the data frames is not necessary, as they will be renamed with
the \code{strategies} vector.}

\item{strategies}{#' Vector with the names of the strategies. Due to requirements in
certain uses of this vector, all spaces in the strategy names will be replaced with an underscore
with a warning.}

\item{cost}{Data frames containing costs and effectiveness data, respectively.
Each simulation should be a row of the data frame, and each strategy should be a column.
Naming the columns of the data frames is not necessary, as they will be renamed with
the \code{strategies} vector.}

\item{currency}{symbol for the currency being used (ex. "$", "£")}
}
\description{
The object returned by this function can be passed to
\code{\link{owsa}} to do a one-way sensitivity analysis on each
parameter of interest.
}
