% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_psa_samp.R
\name{gen_psa_samp}
\alias{gen_psa_samp}
\title{Generate PSA Sample}
\usage{
gen_psa_samp(params = NULL, dist = c("normal", "log-normal", "beta",
  "gamma", "dirichlet", "bootstrap"), parameterization_type = NULL,
  dist_params = NULL, n_samp = 100)
}
\arguments{
\item{params}{String vector with the names of parameters to be generated by \code{gen_psa_samp}
and used by a user-defined function in \code{run_psa} to calculate outcomes.}

\item{dist}{String vector with the distributions from which \code{params} will be drawn.}

\item{parameterization_type}{String vector with parameterization types for each \code{dist}}

\item{dist_params}{list of input parameters required to by specific \code{dist} and \code{parameterization_type}
to fully describe distribution and generate parameter samples.}

\item{n_samp}{number of sets of parameter values to be generated}
}
\value{
A dataframe with samples of parameters for a probabilistic sensitivity analysis (PSA)
}
\description{
This function generates a data.frame of sampled parameter values from user-specified distributions
to be used in a probabilistic sensitivity analysis (PSA)
}
\details{
Fill in details
}
\seealso{
\code{\link{run_psa}},
}
