% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_common_aes.R
\name{labfun}
\alias{labfun}
\title{used to automatically label continuous scales}
\usage{
labfun(x)
}
\arguments{
\item{x}{axis breaks}
}
\value{
a character vector giving a label for each input value
}
\description{
used to automatically label continuous scales
}
\keyword{internal}
