% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{metamod}
\alias{metamod}
\title{linear regression metamodeling}
\usage{
metamod(y, psa, parm, poly.order = 2)
}
\arguments{
\item{y}{data frame with the model outputs of interest}

\item{psa}{psa object}

\item{parm}{String with the name of the parameter of interest}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\description{
linear regression metamodeling
}
