% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{metamod}
\alias{metamod}
\title{linear regression metamodeling}
\usage{
metamod(psa, parms = NULL, strategies = NULL, outcome = c("eff",
  "cost", "nhb", "nmb"), wtp = NULL, poly.order = 2)
}
\arguments{
\item{psa}{psa object}

\item{parms}{String with the name of the parameter of interest}

\item{strategies}{vector of strategies to consider. the default (NULL) is that all strategies are considered. The}

\item{outcome}{either effectiveness ("eff"), cost, net health benefit ("nhb"), or net monetary benefit ("nmb")}

\item{wtp}{if outcome is NHB or NMB, must provide the willingness-to-pay threshold}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\description{
linear regression metamodeling
}
