% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{number_ticks}
\alias{number_ticks}
\title{Number of ticks for \code{ggplot2} plots}
\usage{
number_ticks(n)
}
\arguments{
\item{n}{integer giving the desired number of number of ticks in axis of
\code{ggplot2} plots. Non-integer values are rounded down.}
}
\description{
Function for number of ticks in axis of \code{ggplot2} plots.
}
\section{Details}{

Based on function \code{pretty}.
}
\keyword{ggplot2}

