% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa}
\alias{owsa}
\title{One-way sensitivity analysis using linear regression metamodeling}
\usage{
owsa(psa, parm, newdata = NULL, outcome = c("eff", "cost", "nhb",
  "nmb"), wtp = NULL, strategies = NULL, poly.order = 2)
}
\arguments{
\item{psa}{psa object}

\item{parm}{String with the name of the parameter of interest}

\item{newdata}{data frame with values of parameter of interest}

\item{outcome}{either effectiveness ("eff"), cost, net health benefit ("nhb"), or net monetary benefit ("nmb")}

\item{wtp}{if outcome is NHB or NMB, must provide the willingness-to-pay threshold}

\item{strategies}{vector of strategies to consider. the default (NULL) is that all strategies are considered. The}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\value{
A dataframe with the results of the sensitivity analysis.
Can be visualized with \code{\link{plot.owsa}}
}
\description{
This function displays a one-way sensitivity analysis (OWSA) graph
by estimating a linear regression metamodel of a PSA for a given
decision-analytic model
}
\keyword{analysis;}
\keyword{linear}
\keyword{metamodel}
\keyword{one-way}
\keyword{regression}
\keyword{sensitivity}
