% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa}
\alias{owsa}
\title{One-way sensitivity analysis using linear regression metamodeling}
\usage{
owsa(sa_obj, parms = NULL, ranges = NULL, nsamps = 100,
  outcome = c("eff", "cost", "nhb", "nmb", "nhb_loss", "nmb_loss"),
  wtp = NULL, strategies = NULL, poly.order = 2)
}
\arguments{
\item{sa_obj}{sensitivity analysis object;
either a probabilistic sensitivity analysis (\code{\link{make_psa_obj}}) or
a deterministic sensitivity analysis object (\code{\link{create_dsa_twoway}})}

\item{parms}{String vector with the name(s) of the parameter of interest. Defaults to all.}

\item{ranges}{A named list of the form c("parm" = c(0, 1), ...)
that gives the ranges for the parameter of interest. If NULL,
parameter values from the middle 95\% of the PSA samples are used. The number of samples
from this range is determined by \code{nsamp}.}

\item{nsamps}{number of samples to take from the ranges}

\item{outcome}{either effectiveness ("eff"), cost ("cost"), net health benefit ("nhb"),
net monetary benefit ("nmb"), or the opportunity loss in terms of NHB or
NMB ("nhb_loss" and "nmb_loss", respectively)}

\item{wtp}{if outcome is NHB or NMB (or the associated loss), must provide the willingness-to-pay threshold}

\item{strategies}{vector of strategies to consider. The default (NULL) is that all strategies are considered.}

\item{poly.order}{Order of polynomial for the linear regression metamodel.
Default: 2}
}
\value{
A dataframe with the results of the sensitivity analysis.
Can be visualized with \code{\link{plot.owsa}, \link{owsa_tornado}, and \link{owsa_opt_strat}}
}
\description{
This function uses a linear regression metamodel of a PSA for a given
decision-analytic model to predict the desired outcome.
}
