% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa_opt_strat}
\alias{owsa_opt_strat}
\title{plot the optimal strategy as the parameter values change}
\usage{
owsa_opt_strat(owsa, maximize = TRUE, return = c("plot", "data"),
  plot_const = TRUE, col = c("full", "bw"), greystart = 0.2,
  greyend = 0.8, txtsize = 12, n_x_ticks = 10)
}
\arguments{
\item{owsa}{An owsa object}

\item{maximize}{whether to maximize (TRUE) or minimize the outcome}

\item{return}{either return a ggplot object \code{plot} or a data frame with
ranges of parameters for which each strategy is optimal.}

\item{plot_const}{whether to plot parameters that don't lead to
changes in optimal strategy as they vary.}

\item{col}{either full color or black and white}

\item{greystart}{between 0 and 1. used in greyscale only. smaller numbers are lighter}

\item{greyend}{between 0 and 1, greater than greystart.}

\item{txtsize}{base text size}

\item{n_x_ticks}{number of axis ticks}
}
\description{
plot the optimal strategy as the parameter values change
}
