% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa_opt_strat}
\alias{owsa_opt_strat}
\title{plot the optimal strategy as the parameter values change}
\usage{
owsa_opt_strat(owsa, maximize = TRUE, txtsize = 12,
  plot_const = TRUE, return = c("plot", "data"))
}
\arguments{
\item{owsa}{An owsa object}

\item{maximize}{whether to maximize (TRUE) or minimize the outcome}

\item{txtsize}{base text size for the plot}

\item{plot_const}{whether to plot parameters that don't lead to
changes in optimal strategy as they vary.}

\item{return}{either return a ggplot object \code{plot} or a data frame with
ranges of parameters for which each strategy is optimal.}
}
\description{
plot the optimal strategy as the parameter values change
}
