% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa_tornado}
\alias{owsa_tornado}
\title{Tornado plot of a one-way sensitivity analysis}
\usage{
owsa_tornado(owsa, return = c("plot", "data"), txtsize = 12,
  min_rel_diff = 0, col = c("full", "bw"), n_y_ticks = 8,
  ylim = NULL, ybreaks = NULL)
}
\arguments{
\item{owsa}{an owsa object}

\item{return}{either return a ggplot object \code{plot} or a data frame with
ranges of parameters for which each strategy is optimal.}

\item{txtsize}{base text size}

\item{min_rel_diff}{this function only plots
parameters that lead to a relative change in the outcome greater than or equal
to \code{min_rel_diff}, which must be between 0 and 1. The default (0) is that
no strategies are filtered.}

\item{col}{either none, full color, or black and white}

\item{n_y_ticks}{number of axis ticks}

\item{ylim}{vector of axis limits, or NULL, which sets limits automatically}

\item{ybreaks}{vector of axis breaks.
will override \code{n_x_ticks} and/or \code{n_y_ticks} if provided.}
}
\description{
Tornado plot of a one-way sensitivity analysis
}
