% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{owsa_tornado}
\alias{owsa_tornado}
\title{Tornado plot of a one-way sensitivity analysis}
\usage{
owsa_tornado(owsa, strategy, txtsize = 12, min_rel_diff = 0.01,
  return = c("plot", "data"))
}
\arguments{
\item{owsa}{an owsa object}

\item{strategy}{the desired strategy}

\item{txtsize}{base textsize}

\item{min_rel_diff}{this function only plots
parameters that lead to a relative change in the outcome greater than or equal
to \code{min_rel_diff}, which must be between 0 and 1. To disable filtering, set to 0.}

\item{return}{either return a ggplot object \code{plot} or a data frame with
ranges of parameters for which each strategy is optimal.}
}
\description{
Tornado plot of a one-way sensitivity analysis
}
