% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ceac.R
\name{plot.ceac}
\alias{plot.ceac}
\title{Plot of Cost-Effectiveness Acceptability Curves (CEAC)}
\usage{
\method{plot}{ceac}(x, ..., title = "Cost-Effectiveness Acceptability Curves",
  txtsize = 12, currency = "$")
}
\arguments{
\item{x}{Object of class \code{ceac}. A melted data frame produced by
function \code{ceac} with each strategy's probability of being
cost-effective for each willingness-to-pay (WTP) threshold}

\item{...}{additional arguments to plot (not used)}

\item{title}{String with graph's title}

\item{txtsize}{number with text size}

\item{currency}{String with currency used in the cost-effectiveness analysis (CEA).
Default: $, but it could be any currency symbol or word (e.g., £, €, peso)}
}
\value{
ceac.gg A \code{ggplot2} object with the CEAC
}
\description{
Plots the CEAC as a \code{ggplot2} object calculated with \code{\link{ceac}}.
}
\section{Details}{

\code{ceac} computes the probability of each of the strategies being
cost-effective at each \code{wtp} value.
}
\keyword{acceptability}
\keyword{cost-effectiveness}
\keyword{curves}

