% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi.R
\name{plot.evpi}
\alias{plot.evpi}
\title{Plot of Expected Value of Perfect Information (EVPI)}
\usage{
\method{plot}{evpi}(x, txtsize = 12, currency = "$",
  effect_units = "QALY", n_y_ticks = 8, n_x_ticks = 20,
  xbreaks = NULL, ybreaks = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{object of class \code{evpi}, produced by function
\code{\link{calc_evpi}}}

\item{txtsize}{base text size}

\item{currency}{String with currency used in the cost-effectiveness analysis (CEA).
Default: $, but it could be any currency symbol or word (e.g., £, €, peso)}

\item{effect_units}{Units of effectiveness. Default: QALY}

\item{n_y_ticks}{number of axis ticks}

\item{n_x_ticks}{number of axis ticks}

\item{xbreaks}{vector of axis breaks.
will override \code{n_x_ticks} and/or \code{n_y_ticks} if provided.}

\item{ybreaks}{vector of axis breaks.
will override \code{n_x_ticks} and/or \code{n_y_ticks} if provided.}

\item{xlim}{vector of axis limits, or NULL, which sets limits automatically}

\item{ylim}{vector of axis limits, or NULL, which sets limits automatically}

\item{...}{further arguments to plot (not used)}
}
\value{
A \code{ggplot2} object with the EVPI
}
\description{
Plots the EVPI as a \code{ggplot2} object calculated with \code{\link{calc_evpi}}.
}
\keyword{expected}
\keyword{information}
\keyword{of}
\keyword{perfect}
\keyword{value}
