% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.icers.R
\name{plot.icers}
\alias{plot.icers}
\title{Plot of ICERs}
\usage{
\method{plot}{icers}(x, ..., frontier = TRUE, title = "",
  txtsize = 12, currency = "$", effect_units = "QALYs",
  label = FALSE)
}
\arguments{
\item{x}{Object of class \code{ceac}. A melted data frame produced by
function \code{ceac} with each strategy's probability of being
cost-effective for each willingness-to-pay (WTP) threshold}

\item{...}{additional arguments to plot (not used)}

\item{frontier}{Whether to plot acceptability frontier}

\item{title}{String with graph's title}

\item{txtsize}{integer. base font size}

\item{currency}{string. with currency used in the cost-effectiveness analysis (CEA).}

\item{effect_units}{string. unit of effectiveness}

\item{label}{boolean. whether or not to label the strategies.
may not be that useful when strategies have long names. Defaults to FALSE.

todo: type checking.}
}
\description{
Plots the CEAC as a \code{ggplot2} object calculated with \code{\link{ceac}}.
}
