% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owsa.R
\name{plot.owsa}
\alias{plot.owsa}
\title{Plot a sensitivity analysis}
\usage{
\method{plot}{owsa}(x, txtsize = 12, ...)
}
\arguments{
\item{x}{an owsa object}

\item{txtsize}{base text size in the plot}

\item{...}{further arguments to \code{plot.owsa} (not used)}
}
\description{
Plot a sensitivity analysis
}
