% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psa.R
\name{plot.psa}
\alias{plot.psa}
\title{Plot the psa object}
\usage{
\method{plot}{psa}(x, ..., center = TRUE, ellipse = TRUE)
}
\arguments{
\item{x}{the psa object}

\item{...}{further arguments to plot (not used)}

\item{center}{plot the mean cost and effectiveness for each strategy. defaults to TRUE}

\item{ellipse}{plot an ellipse around each strategy. defaults to TRUE}
}
\description{
Plot the psa object
}
