% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa.R
\name{plot.twsa}
\alias{plot.twsa}
\title{Two-way sensitivity analysis plot}
\usage{
\method{plot}{twsa}(x, maximize = TRUE, col = c("full", "bw"),
  n_x_ticks = 6, n_y_ticks = 6, txtsize = 12, ...)
}
\arguments{
\item{x}{a twsa object}

\item{maximize}{If \code{TRUE}, plot of strategy with maximum expected outcome
(default); if \code{FALSE}, plot of strategy with minimum expected outcome}

\item{col}{either none, full color, or black and white}

\item{n_x_ticks}{number of axis ticks}

\item{n_y_ticks}{number of axis ticks}

\item{txtsize}{base text size}

\item{...}{further arguments to plot.
This is not used by \code{dampack} but required for generic consistency.}
}
\description{
Two-way sensitivity analysis plot
}
