% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa.R
\name{plot.twsa}
\alias{plot.twsa}
\title{Two-way sensitivity analysis plot}
\usage{
\method{plot}{twsa}(x, txtsize = 12, maximize = TRUE, title = NULL,
  ...)
}
\arguments{
\item{x}{a twsa object}

\item{txtsize}{base text size in the plot}

\item{maximize}{If \code{TRUE}, plot of strategy with maximum expected outcome
(default); if \code{FALSE}, plot of strategy with minimum expected outcome}

\item{title}{Title for the plot}

\item{...}{further arguments to \code{plot.owsa} (not used)}
}
\description{
Two-way sensitivity analysis plot
}
