% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{predict.metamodel}
\alias{predict.metamodel}
\title{Predict from a metamodel}
\usage{
\method{predict}{metamodel}(object, newdata = NULL, nsamp = 400, ...)
}
\arguments{
\item{object}{object with class "metamodel"}

\item{newdata}{values for parameter of interest. if NULL,
parameter values from the middle 95% are used. The number of samples
from this range is determined by \code{nsamp}}

\item{nsamp}{number of samples from range}

\item{...}{further arguments to \code{predict} (not used)}
}
\description{
Predict from a metamodel
}
