% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{predict.metamodel}
\alias{predict.metamodel}
\title{Predict from a one-way or two-way metamodel}
\usage{
\method{predict}{metamodel}(object, ranges = NULL, nsamp = 100, ...)
}
\arguments{
\item{object}{object with class "metamodel"}

\item{ranges}{a named list of the form c("param" = c(0, 1), ...)
that gives the ranges for the parameter of interest. If NULL,
parameter values from the middle 95% of the PSA samples are used. The number of samples
from this range is determined by \code{nsamp}.}

\item{nsamp}{number of samples from ranges}

\item{...}{further arguments to \code{predict} (not used)}
}
\value{
a \code{data.frame} containing the outcome values predicted by the metamodel
for each strategy and each combination of parameter values defined by \code{ranges}.
}
\description{
Predict from a one-way or two-way metamodel
}
