% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icers.R
\name{print.icers}
\alias{print.icers}
\title{Print the formatted CEA table}
\usage{
\method{print}{icers}(x, currency = "$", dig_cost = 0, dig_eff = 2, dig_icer = 0, ...)
}
\arguments{
\item{x}{Object of class \code{icers}}

\item{currency}{string. with currency used in the cost-effectiveness analysis (CEA).}

\item{dig_cost}{integer. number of digits to round costs to}

\item{dig_eff}{integer. number of digits to round effects to}

\item{dig_icer}{integer. number of digits to round ICER to}
}
\value{
a dataframe object - formatted CEA table
}
\description{
Print the formatted CEA table
}
