% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{print.metamodel}
\alias{print.metamodel}
\title{Print metamodel}
\usage{
\method{print}{metamodel}(x, ...)
}
\arguments{
\item{x}{metamodel to print}

\item{...}{further arguments to print}
}
\description{
Print metamodel
}
