% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_psa_samp.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random number generation for the Dirichlet distribution with parameter vector alpha.}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of observations}

\item{alpha}{vector of parameters defining Dirichlet distribution


 @importFrom stats rgamma
 @return A vector random values sampled from a dirichlet distribution
 @export}
}
\description{
Random number generation for the Dirichlet distribution with parameter vector alpha.
}
