% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dsa.R
\name{run_owsa_det}
\alias{run_owsa_det}
\title{Run deterministic one-way sensitivity analysis (OWSA)}
\usage{
run_owsa_det(params = NULL, params_all, nsamp = 100, FUN,
  outcomes = NULL, strategies = NULL, ...)
}
\arguments{
\item{params}{Vector with strings with the name of the parameters of interest}

\item{params_all}{A data.frame with 4 columns with following column order: "pars",
"basecase", "min", and "max". The number of samples from this range is
determined by \code{nsamp}}

\item{nsamp}{number of parameter values. If \code{NULL}, 100 parameter values are
used}

\item{FUN}{Function that takes the basecase in \code{params_all} and \code{...} to
produce the \code{outcome} of interest. The \code{FUN} must return a dataframe
where the first column are the strategy names and the rest of the columns must be outcomes.}

\item{outcomes}{String vector with the outcomes of interest from \code{FUN}
produced by \code{nsamp}}

\item{strategies}{vector of strategy names. The default \code{NULL} will use
strategy names in \code{FUN}}

\item{...}{Additional arguments to user-defined \code{FUN}}
}
\value{
A list containing dataframes with the results of the sensitivity analyses.
The list will contain a dataframe for each outcome specified.
List elements can be visualized with \code{plot.owsa},
\code{owsa_opt_strat} and
\code{owsa_tornado} from \code{dampack}
}
\description{
This function runs a deterministic one-way sensitivity analysis (OWSA) on a
given function that produces outcomes.
}
\section{Details}{

\itemize{
\item \code{params_all}
\itemize{
\item "pars" are the names of the input parameters in the
user defined function. "pars" should include all parameters of interest provided in
\code{params}.
\item "basecase" are the base value of input parameters in user defined \code{FUN}.
\item "min" and "max" are the mininum and maximum values of the parameters of interest.
Users only need to provide the "min" and "max" of the parameters of interest.
For the rest of the parameter inputs into the user defined function,
"min" and "max" can be any value or \code{NA} but these values are not evaluated in
\code{run_owsa_det}}
}
}

