% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{run_psa}
\alias{run_psa}
\title{Calculate outcomes for a PSA using a user-defined function.}
\usage{
run_psa(psa_samp, FUN, outcomes = NULL, cost_outcome = NULL,
  effectiveness_outcome = NULL, strategies = NULL, currency = "$",
  ...)
}
\arguments{
\item{psa_samp}{A dataframe with samples of parameters for a probabilistic sensitivity analysis (PSA)}

\item{FUN}{Function that takes the parameter values in \code{psa_samp} and \code{...} to
produce the \code{outcome} of interest. The \code{FUN} must return a dataframe
where the first column are the strategy names and the rest of the columns must be outcomes.}

\item{outcomes}{String vector with the outcomes of interest from \code{FUN}.}

\item{cost_outcome}{String within \code{outcomes} that is designated as the cost outcome
in the creation of \code{psa} objects for use in cost-effectiveness analyses.}

\item{effectiveness_outcome}{String within \code{outcomes} that is designated as the effectiveness outcome
in the creation of \code{psa} objects for use in cost-effectiveness analyses.}

\item{strategies}{vector of strategy names. The default \code{NULL} will use
strategy names in \code{FUN}}

\item{currency}{symbol for the currency being used (ex. "$", "£")}

\item{...}{Additional arguments to user-defined \code{FUN}}
}
\value{
A list containing PSA objects for each outcome in \code{outcomes},
as well as an additional PSA object containing both effectiveness and cost for CEA
if cost_outcome and effectiveness_outcome were both supplied.
}
\description{
\code{run_psa} calculates outcomes using a user-defined function and creates PSA objects
corresponding to the specified outcomes.
}
\details{
Fill in details
}
\seealso{
\code{\link{run_psa}},
\code{\link{make_psa_obj}},
\code{\link{gen_psa_samp}},
}
