% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{summary.ceac}
\alias{summary.ceac}
\title{Summarize a ceac}
\usage{
\method{summary}{ceac}(object, ...)
}
\arguments{
\item{object}{object returned from the \code{ceac} function}

\item{...}{further arguments (not used)}
}
\value{
data frame showing the interval of cost effectiveness for each
interval. The intervals are open on the right endpoint -
i.e., [\code{range_min}, \code{range_max})
}
\description{
Describes cost-effective strategies and their
associated intervals of cost-effectiveness
}
\keyword{internal}
