% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ceac.R
\name{summary.ceac}
\alias{summary.ceac}
\title{summary function for the ceac object}
\usage{
\method{summary}{ceac}(object, ...)
}
\arguments{
\item{object}{object returned from the \code{ceac} function}

\item{...}{further arguments (not used)}
}
\description{
summary function for the ceac object
}
