% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamod.R
\name{summary.metamodel}
\alias{summary.metamodel}
\title{Summary of metamodel}
\usage{
\method{summary}{metamodel}(object, ...)
}
\arguments{
\item{object}{metamodel to summarize}

\item{...}{further arguments to summary}
}
\description{
Summary of metamodel
}
