% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.psa.R
\name{summary.psa}
\alias{summary.psa}
\title{summarize a psa object across all simulations}
\usage{
\method{summary}{psa}(object, ..., calc_sds = FALSE)
}
\arguments{
\item{object}{the psa object}

\item{...}{further arguments to summary (not used)}

\item{calc_sds}{whether or not to calculate the standard deviations. Defaults to FALSE}
}
\description{
summarize a psa object across all simulations
}
