% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic-SA.R
\name{twsa_det}
\alias{twsa_det}
\title{Two-way sensitivity analysis (TWSA)}
\usage{
twsa_det(parm1, parm2, pars_df, nsamps = 40, FUN, outcome,
  outcome_type = "eff", strategies = NULL, ...)
}
\arguments{
\item{parm1}{String with the name of the first parameter of interest}

\item{parm2}{String with the name of the second parameter of interest}

\item{pars_df}{A data.frame with 4 columns with following column order: "pars",
"basecase", "min", and "max". The number of samples from this range is
determined by \code{nsamp}}

\item{nsamps}{number of parameter values. If \code{NULL}, 40 parameter values are
used}

\item{FUN}{Function that takes the basecase in \code{pars_df} and \code{...} to
produce the \code{outcome} of interest}

\item{outcome}{String with the outcome of interest produced by \code{nsamp}}

\item{outcome_type}{The type of outcome is either "eff" or "cost". The default
outcome_type is "eff"}

\item{strategies}{vector of strategy names. The default (NULL) will use
strategy names in FUN}

\item{...}{Additional arguments to user-defined \code{FUN}}
}
\value{
A dataframe with the results of the sensitivity analysis. Can be
visualized with ??
}
\description{
This function runs a deterministic two-way sensitivity analysis (TWSA) on a
given function that produces outcomes.
}
\section{Details}{

\itemize{
\item \code{pars_df}
\itemize{
\item "pars" are the names of the input parameters in the
user defined function. "pars" should include all parameters of interest provided in
\code{parms}.
\item "basecase" are the base value of input parameters in user defined \code{FUN}.
\item "min" and "max" are the mininum and maximum values of the parameters of interest.
Users only need to provide the "min" and "max" of the parameters of interest.
For the rest of the parameter inputs into the user defined function,
"min" and "max" can be any value or \code{NA} but these values are not evaluated in
\code{twsa_det}}
\item \code{FUN} must return a dataframe where the first column are the strategy names
and the rest of the columns must be outcomes.
}
}

\keyword{twsa}
